Improved logging and fix for procedures blocking forever in apoc.periodic.iterate
The fix is based on the discovery that nodes produced by the driving table are not re-created within the inner transactions, and therefor not safe to use. We fix this by:

    node = tx.getNodeById(node.getId());

We had assumed that apoc.periodic.iterate did this for all objects that contained handles to previous transactions, like Node and Relationship, but this turns out not to be true.